+++ lib/emailbutler/mailers/helpers.rb
@@ -18,7 +18,7 @@ module Emailbutler
       end
 
       def create_emailbutler_message(args)
-        @message = ::Emailbutler::Message.new(
+        @message = Emailbutler.build(
           status: 'created',
           mailer: self.class.to_s,
           action: action_name,
@@ -35,8 +35,8 @@ module Emailbutler
       end
 
       def set_send_to_for_message
-        @message.send_to = mail.to
-        @message.save
+        Emailbutler.set_attribute(@message, :send_to, mail.to)
+        Emailbutler.save(@message)
       end
     end
   end