+++ spec/dummy/config/initializers/emailbutler.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+require 'emailbutler/adapters/active_record'
+
+Emailbutler.configure do |config|
+  config.adapter = Emailbutler::Adapters::ActiveRecord.new
+end