+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - saving additional mail params to message
 - Webhooks::Receiver class for checking webhooks payload and updating messages
 - message params serialization with global id support
+- DSL to all the emailbutler goodness
+- active record adapter for DSL
 
 ## [0.1.0] - 2022-09-01
 ### Added